# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: We Sell Houses API Documentation
  description: API for property management and communication between user and property agent
  contact:
    name: Filip Prosovsky
    email: prosovskyf@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://observe-prime.codio-box.uk/api/v1'
    description: Development server for this application api
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
    /user/profile:
      summary: User management
      description: This resource manages user profile and avatar picture
      get:
        tags:
          - Agents
        description:
          Get user profile based on agentid in JWT token
        security:
          - BearerAuth: []
        responses:
          '200':
            description: View of user profile
            content:
              application/json:
                schema:
                  $ref: ./agents.json#/definitions/agentProfile
          '400':
            description: Error for bad request
          '403':
            description: Not authorized to view this profile
          '404':
            description: Not found
    /user/updateinfo:
      put:
        tags:
          - Agents
        description:
          Update user based on agentid in JWT token
        security:
          - BearerAuth: []
        requestBody:
          description: Updated information
          required: true
          content:
            application/json:
              schema:
                $ref: ./agents.json#/definitions/agentUpdate
        responses:
          '201':
            description: Updated information sent back
            content:
              application/json:
                schema:
                  $ref: ./agents.json#/definitions/agentUpdate
          '400':
            description: Wrong values detected and catched in error
          '403':
            description: Not authorized to view this profile
    /user/changepass:
      put:
        tags:
          - Agents
        description:
          Change user password
        security:
          - BearerAuth: []
        requestBody:
          description: Update information
          required: true
          content:
            application/json:
              schema:
                type: string
                properties:
                  secret:
                    type: string
                required:
                  - secret
        responses:
          '200':
            description: Changed password
          '400':
            description: Not fulfiled minumum requirements for password
          '403':
            description: Wrong current password or not authorized
    /user/avatar:
      put:
        tags:
          - Agents
        description:
          Upload user avatar based on agentid in JWT token
        security:
          - BearerAuth: []
        requestBody:
          description: Picture
          required: true
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  avatar:
                    type: object
                    items:
                      type: string
                      format: binary
                required:
                  - avatar
        responses:
          '200':
            description: Picture uploaded, return picture path
          '403':
            description: Not authorized
      delete:
        tags:
          - Agents
        description:
          Delete user avatar based on agentid in JWT token
        security:
          - BearerAuth: []
        responses:
          '200':
            description: Picture deleted
          '403':
            description: Not authorized
...
    /category:
      summary: Categories
      description: This resource manages categories
      get:
        tags:
          - Categories
        description:
          Get all categories
        responses:
          '200':
            description: Object with categories in array
            content:
              application/json:
                schema:
                  $ref: ./categories.json#/definitions/categoryView
          '404':
            description: Not found
      post:
        tags:
          - Categories
        description:
          Create category
        security:
          - BearerAuth: []
        requestBody:
          description: Category information
          required: true
          content:
            application/json:
              schema:
                $ref: ./categories.json#/definitions/category
        responses:
          '201':
            description: ID of created category
            content:
              application/json:
                schema:
                  type: integer
          '400':
            description: Wrong values entered
          '403':
            description: Not authorized
...
    /features:
      summary: Features
      description: This resource manages features
      get:
        tags:
          - Features
        description:
          Get all features
        responses:
          '200':
            description: Features in array
            content:
              application/json:
                schema:
                  $ref: ./features.json#/definitions/featureView
          '404':
            description: Not found
    /features/{id}:
      post:
        parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the property
        tags:
          - Features
        description:
          Create feature
        security:
          - BearerAuth: []
        requestBody:
          description: Feature data
          required: true
          content:
            application/json:
              schema:
                $ref: ./features.json#/definitions/feature
        responses:
          '201':
            description: Feature created
          '400':
            description: Wrong values entered
          '403':
            description: Not authorized
      delete:
        parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the feature
        tags:
          - Features
        description:
          Delete specific feature
        security:
          - BearerAuth: []
        responses:
          '201':
            description: Feature deleted
          '400':
            description: Wrong ID input
          '403':
            description: Not authorized
...